backend/requirements.txt
# Framework de Agentes
crewai
langchain
langchain-openai
openai

# API Framework
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# OCR
pytesseract
pdf2image
opencv-python
Pillow
PyPDF2

# NLP
spacy
transformers

# Data Processing
pandas
numpy
openpyxl

# Database
psycopg2-binary
pymongo
redis
sqlalchemy

# HTTP Clients
httpx
requests

# Utilities
python-dateutil
python-dotenv
pyyaml
jinja2

# Testing
pytest
pytest-asyncio
